create position beck

+++ server/controllers/analyticalCategoryController.js
@@ -8,21 +8,12 @@ import { generateRefreshToken } from "../utils/generateRefreshToken.js";
 // GET: /api/analyticalcategory/get
 export const getAllCategoryController = async (req, res) => {
     try {
-        const data = await AnalyticalCategory.find({}).sort({ createdAt: -1 });
-        res.status(200).json({
-            success: true,
-            error: false,
-            message: "Каталоги успішно завантажені",
-            data
-        })
-    } catch (error) {
-        console.log(error.message || error);
 
-        res.status(500).json({
-            success: false,
-            error: true,
-            message: error.message || error
-        })
+        const data = await AnalyticalCategory.find({}).sort({ createdAt: -1 });
+        res.status(200).json({ success: true, error: false, message: "Каталоги успішно завантажені", data })
+    } catch (error) {
+        console.log("getAllCategoryController:", error);
+        res.status(500).json({ success: false, error: true, message: error.message || error })
     }
 };
 
@@ -30,38 +21,27 @@ export const getAllCategoryController = async (req, res) => {
 // POST: /api/analyticalcategory/add
 export const addCategoryController = async (req, res) => {
     try {
-        const category = req.body
-        const { userId } = req
-        const _id = userId
-
-        const user = await User.findById(_id)
+        const { userId } = req;
+        const user = await User.findById(userId);
         if (!user) {
-            return res.status(404).json({
-                success: false,
-                error: true,
-                message: "User not found"
-            })
+            return res.status(401).json({ success: false, message: "Користувача не знайдено" });
         }
         if (!user.verify_email) {
-            return res.status(403).json({
-                message: "Підтвердіть email перед входом",
-            });
+            return res.status(403).json({ message: "Підтвердіть email перед входом", succses: false, error: true });
         }
 
-        const slug = slugify(category.title, {
-            lower: true,
-            strict: true,
-            locale: 'uk' // Важно для корректной транслитерации кириллицы
-        });
+        const category = req.body
+        if (!category.title || !category.description) {
+            return res.status(400).json({ success: false, message: "Дані відсутні" });
+        }
 
-        await AnalyticalCategory.create({ owner: _id, title: category.title, description: category.description, slug: slug });
+        const slug = slugify(category.title, { lower: true, strict: true, locale: 'uk' });
+        await AnalyticalCategory.create({ owner: user._id, title: category.title, description: category.description, slug: slug });
 
         const token = generateSessionToken(user._id);
-        const refresh_token = generateRefreshToken(user._id);
-
-        res.status(200).json({ success: true, error: false, token, refresh_token, message: "Категорія успішно додана", })
+        res.status(200).json({ success: true, error: false, token, message: "Категорія успішно додана", })
     } catch (error) {
-        console.log(error.message || error);
+        console.log("addCategoryController:", error);
         res.status(500).json({ success: false, error: true, message: error.message || error })
     }
 };
@@ -70,32 +50,26 @@ export const addCategoryController = async (req, res) => {
 // POST: /api/analyticalcategory/delete
 export const deleteCategoryController = async (req, res) => {
     try {
-        const { id } = req.body;
-        const { userId } = req
-        const _id = userId
-        const user = await User.findById(_id)
-
+        const { userId } = req;
+        const user = await User.findById(userId);
         if (!user) {
-            return res.status(404).json({
-                success: false,
-                error: true,
-                message: "User not found"
-            })
+            return res.status(401).json({ success: false, message: "Користувача не знайдено" });
         }
         if (!user.verify_email) {
-            return res.status(403).json({
-                message: "Підтвердіть email перед входом",
-            });
+            return res.status(403).json({ message: "Підтвердіть email перед входом", succses: false, error: true });
+        }
+
+        const { id } = req.body;
+        if (!id) {
+            return res.status(400).json({ success: false, message: "Дані відсутні" });
         }
 
         await AnalyticalCategory.findByIdAndDelete(id);
 
         const token = generateSessionToken(user._id);
-        const refresh_token = generateRefreshToken(user._id);
-
-        res.status(200).json({ success: true, error: false, token, refresh_token, message: "Категорія успішно видалена", })
+        res.status(200).json({ success: true, error: false, token, message: "Категорія успішно видалена", })
     } catch (error) {
-        console.log(error.message || error);
+        console.log("deleteCategoryController:", error);
         res.status(500).json({ success: false, error: true, message: error.message || error })
     }
 };
